requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
joblib>=1.3.0
pyyaml>=6.0

# Backend
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
openai>=1.40.0
python-dotenv>=1.0.0

# Smoke test / HTTP client
requests>=2.31.0

# Linting & Formatting
flake8>=7.0.0
black>=24.0.0
isort>=5.13.0
pre-commit>=3.6.0

# Testing
pytest>=7.4.0
